Add better ARRAYSIZE, change schedule to use size_t, cleanup
Added better ARRAYSIZE that detects pointers
Changed Schedule_t to use size_t for array size
Cleaned up TODO items, client side model loading

+++ game/server/entities/CBreakable.cpp
@@ -104,7 +104,7 @@ void CBreakable::KeyValue( KeyValueData* pkvd )
 	else if (FStrEq(pkvd->szKeyName, "spawnobject") )
 	{
 		int object = atoi( pkvd->szValue );
-		if ( object > 0 && object < ARRAYSIZE(pSpawnObjects) )
+		if ( object > 0 && static_cast<size_t>( object ) < ARRAYSIZE(pSpawnObjects) )
 			m_iszSpawnObject = MAKE_STRING( pSpawnObjects[object] );
 		pkvd->fHandled = true;
 	}
